Remove utils/path
